src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/iframe-shadow-expected.txt

CONSOLE ERROR: line 33: Uncaught TypeError: host.createShadowRoot is not a function
This test ensures that iframe inside shadow isn't visible from the host document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.length is 4
PASS i0byname.length is 1
PASS i0 is i0byname[0].contentWindow
PASS i1.length is 2
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

   
